logic re-arrangment
De-coupled the wrong next number and two messages from the same user checks so that users could not exploit the bad input try-catch to skip the same user check. As a matter of course I also cleaned up the part where you fetched the last couple messages from the channel, this could be further optimized, but w.e.